requirements.txt
Django==3.1
psycopg2-binary==2.8.5
Pillow==7.1
djangorestframework==3.11
djoser==2.0.3
django-filter
drf-yasg
python-decouple
celery